requirements.txt
# Core agent dependencies
bee-agent==0.2.0
redis==6.1.0
python-dotenv==1.1.0
pydantic==2.11.5
pydantic-core==2.33.2
pydantic-settings==2.9.1
httpx>=0.24.0
aiohttp>=3.8.0

# FastAPI dependencies
fastapi>=0.104.0
uvicorn>=0.24.0